standardizing of structure, Logging to log/application.logs, resonable naming
